feat(db): consolidate duplicate get_db() implementations into a single cannonical function and update imports
