+++ app/api/deps/common.py
@@ -2,7 +2,7 @@ from fastapi import Depends, Header, HTTPException
 from sqlalchemy.orm import Session
 
 from app.core.settings import settings
-from app.infrastructure.db.session import get_db
+from app.api.deps.db import get_db
 
 
 def get_db_session(db: Session = Depends(get_db)) -> Session:
